Update Tasks naming convention, add getter functions to respective class type
Todos Events Deadlines class name in plural no getter function for respective class type

Update class name to singular add getter function for detail, date, mark for respective class types
