+++ src/main/java/tasks/Event.java
@@ -1,11 +1,21 @@
 package tasks;
 
-public class Events extends Task {
+public class Event extends Task {
 	private String date;
 
-	public Events(String detail, String date) {
+	public Event(String detail, String date) {
 		super(detail);
 		this.date = date;
+	}
+
+	@Override
+	public String getDate(){
+		return date;
+	}
+
+	@Override
+	public String getType(){
+		return "E";
 	}
 
 	@Override
